requirements.txt
notebook
numpy
matplotlib
Pillow
pycocotools